Fix broken contact groups
When adding a contact to a group both the group and contact must be created or referenced with the same address book (ABAddressBookCreate()).

Otherwise ABGroup.addMember fails silently with no error.

Revised the code to create one address book and re-use it.  This required replacing the class methods to instance methods.