Fixed potential crash bug
ABMultivalueCopyLabelAtIndex may return null. Null values cannot be
added to an NSMutableArray. This patch tries to detect if the label
string is null before adding it to the NSMutableArray. Instead of
trying to add the null value and crashing as a result, an empty string
is added instead. The string could easily be something like "Other" if
you plan on using the labels but for my purposes an empty string works
just fine.